Move SurrogateBuilder impl in gpmix
