+++ ego/src/solver/egor_impl.rs
@@ -8,12 +8,10 @@ use crate::{EgorSolver, DEFAULT_CSTR_TOL};
 use argmin::core::{CostFunction, Problem};
 use egobox_doe::{Lhs, LhsKind, SamplingMethod};
 use egobox_gp::ThetaTuning;
-use egobox_moe::{
-    Clustering, CorrelationSpec, GpMixtureParams, MixtureGpSurrogate, RegressionSpec,
-};
 use env_logger::{Builder, Env};
 use finitediff::FiniteDiff;
-use linfa::ParamGuard;
+
+use egobox_moe::{Clustering, MixtureGpSurrogate};
 use log::{debug, info, warn};
 use ndarray::{
     concatenate, s, Array, Array1, Array2, ArrayBase, ArrayView2, Axis, Data, Ix1, Ix2, Zip,
@@ -21,63 +19,6 @@ use ndarray::{
 use ndarray_stats::QuantileExt;
 use rand_xoshiro::Xoshiro256Plus;
 use rayon::prelude::*;
-
-impl SurrogateBuilder for GpMixtureParams<f64> {
-    /// Constructor from domain space specified with types
-    /// **panic** if xtypes contains other types than continuous type `Float`
-    fn new_with_xtypes(xtypes: &[XType]) -> Self {
-        if crate::utils::discrete(xtypes) {
-            panic!("GpMixtureParams cannot be created with discrete types!");
-        }
-        GpMixtureParams::new()
-    }
-
-    /// Sets the allowed regression models used in gaussian processes.
-    fn set_regression_spec(&mut self, regression_spec: RegressionSpec) {
-        *self = self.clone().regression_spec(regression_spec);
-    }
-
-    /// Sets the allowed correlation models used in gaussian processes.
-    fn set_correlation_spec(&mut self, correlation_spec: CorrelationSpec) {
-        *self = self.clone().correlation_spec(correlation_spec);
-    }
-
-    /// Sets the number of components to be used specifiying PLS projection is used (a.k.a KPLS method).
-    fn set_kpls_dim(&mut self, kpls_dim: Option<usize>) {
-        *self = self.clone().kpls_dim(kpls_dim);
-    }
-
-    /// Sets the number of clusters used by the mixture of surrogate experts.
-    fn set_n_clusters(&mut self, n_clusters: usize) {
-        *self = self.clone().n_clusters(n_clusters);
-    }
-
-    /// Sets the number of clusters used by the mixture of surrogate experts.
-    fn set_theta_tunings(&mut self, theta_tunings: &[ThetaTuning<f64>]) {
-        *self = self.clone().theta_tunings(theta_tunings);
-    }
-
-    fn train(
-        &self,
-        xt: &ArrayView2<f64>,
-        yt: &ArrayView2<f64>,
-    ) -> Result<Box<dyn MixtureGpSurrogate>> {
-        let checked = self.check_ref()?;
-        let moe = checked.train(xt, yt)?;
-        Ok(moe).map(|moe| Box::new(moe) as Box<dyn MixtureGpSurrogate>)
-    }
-
-    fn train_on_clusters(
-        &self,
-        xt: &ArrayView2<f64>,
-        yt: &ArrayView2<f64>,
-        clustering: &Clustering,
-    ) -> Result<Box<dyn MixtureGpSurrogate>> {
-        let checked = self.check_ref()?;
-        let moe = checked.train_on_clusters(xt, yt, clustering)?;
-        Ok(moe).map(|moe| Box::new(moe) as Box<dyn MixtureGpSurrogate>)
-    }
-}
 
 impl<SB: SurrogateBuilder> EgorSolver<SB> {
     /// Constructor of the optimization of the function `f` with specified random generator
